Admin can now create scheduled work shifts for a user.
These will include scheduled in times and out times as well as a location.

If a user is not assigned, it is assumed that the shift is an open shift that other staff can claim.

The schedule view now shows all scheduled shifts for a certain day. Will fix some css issues in future commits.