Admin can now create a shifts on a day per day basis.
Leaving a shift unassigned to a user is assumed to be an available shift that staff can pick up.
When viewing the scheduling tab, admin has options to choose from, such as viewing scheduled shifts and available ones.
The schedule home page also displays how many people are available for time periods.